LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/transform-rotate-composition-expected.txt


FAIL Compositing: property <transform> underlying [rotateX(20deg)] from add [rotateX(40deg)] to add [rotateX(60deg)] at (-0.5) should be [rotateX(50deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.64 , 0.77 , 0 , 0 , - 0.77 , 0.64 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.87 , 0.5 , 0 , 0 , - 0.5 , 0.87 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(20deg)] from add [rotateX(40deg)] to add [rotateX(60deg)] at (0) should be [rotateX(60deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.5 , 0.87 , 0 , 0 , - 0.87 , 0.5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.77 , 0.64 , 0 , 0 , - 0.64 , 0.77 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(20deg)] from add [rotateX(40deg)] to add [rotateX(60deg)] at (0.25) should be [rotateX(65deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.42 , 0.91 , 0 , 0 , - 0.91 , 0.42 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , - 0.71 , 0.71 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(20deg)] from add [rotateX(40deg)] to add [rotateX(60deg)] at (0.5) should be [rotateX(70deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.34 , 0.94 , 0 , 0 , - 0.94 , 0.34 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.64 , 0.77 , 0 , 0 , - 0.77 , 0.64 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(20deg)] from add [rotateX(40deg)] to add [rotateX(60deg)] at (0.75) should be [rotateX(75deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.26 , 0.97 , 0 , 0 , - 0.97 , 0.26 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.57 , 0.82 , 0 , 0 , - 0.82 , 0.57 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(20deg)] from add [rotateX(40deg)] to add [rotateX(60deg)] at (1) should be [rotateX(80deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.17 , 0.98 , 0 , 0 , - 0.98 , 0.17 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.5 , 0.87 , 0 , 0 , - 0.87 , 0.5 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(20deg)] from add [rotateX(40deg)] to add [rotateX(60deg)] at (1.5) should be [rotateX(90deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , - 1 , 0 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.34 , 0.94 , 0 , 0 , - 0.94 , 0.34 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from add [rotateY(40deg)] to add [rotateY(60deg)] at (-0.5) should be [rotateY(50deg)] assert_equals: expected "matrix3d ( 0.64 , 0 , - 0.77 , 0 , 0 , 1 , 0 , 0 , 0.77 , 0 , 0.64 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.87 , 0 , - 0.5 , 0 , 0 , 1 , 0 , 0 , 0.5 , 0 , 0.87 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from add [rotateY(40deg)] to add [rotateY(60deg)] at (0) should be [rotateY(60deg)] assert_equals: expected "matrix3d ( 0.5 , 0 , - 0.87 , 0 , 0 , 1 , 0 , 0 , 0.87 , 0 , 0.5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.77 , 0 , - 0.64 , 0 , 0 , 1 , 0 , 0 , 0.64 , 0 , 0.77 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from add [rotateY(40deg)] to add [rotateY(60deg)] at (0.25) should be [rotateY(65deg)] assert_equals: expected "matrix3d ( 0.42 , 0 , - 0.91 , 0 , 0 , 1 , 0 , 0 , 0.91 , 0 , 0.42 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.71 , 0 , - 0.71 , 0 , 0 , 1 , 0 , 0 , 0.71 , 0 , 0.71 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from add [rotateY(40deg)] to add [rotateY(60deg)] at (0.5) should be [rotateY(70deg)] assert_equals: expected "matrix3d ( 0.34 , 0 , - 0.94 , 0 , 0 , 1 , 0 , 0 , 0.94 , 0 , 0.34 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.64 , 0 , - 0.77 , 0 , 0 , 1 , 0 , 0 , 0.77 , 0 , 0.64 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from add [rotateY(40deg)] to add [rotateY(60deg)] at (0.75) should be [rotateY(75deg)] assert_equals: expected "matrix3d ( 0.26 , 0 , - 0.97 , 0 , 0 , 1 , 0 , 0 , 0.97 , 0 , 0.26 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.57 , 0 , - 0.82 , 0 , 0 , 1 , 0 , 0 , 0.82 , 0 , 0.57 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from add [rotateY(40deg)] to add [rotateY(60deg)] at (1) should be [rotateY(80deg)] assert_equals: expected "matrix3d ( 0.17 , 0 , - 0.98 , 0 , 0 , 1 , 0 , 0 , 0.98 , 0 , 0.17 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.5 , 0 , - 0.87 , 0 , 0 , 1 , 0 , 0 , 0.87 , 0 , 0.5 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from add [rotateY(40deg)] to add [rotateY(60deg)] at (1.5) should be [rotateY(90deg)] assert_equals: expected "matrix3d ( 0 , 0 , - 1 , 0 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.34 , 0 , - 0.94 , 0 , 0 , 1 , 0 , 0 , 0.94 , 0 , 0.34 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from add [rotateZ(40deg)] to add [rotateZ(60deg)] at (-0.5) should be [rotateZ(50deg)] assert_equals: expected "matrix ( 0.64 , 0.77 , - 0.77 , 0.64 , 0 , 0 ) " but got "matrix ( 0.87 , 0.5 , - 0.5 , 0.87 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from add [rotateZ(40deg)] to add [rotateZ(60deg)] at (0) should be [rotateZ(60deg)] assert_equals: expected "matrix ( 0.5 , 0.87 , - 0.87 , 0.5 , 0 , 0 ) " but got "matrix ( 0.77 , 0.64 , - 0.64 , 0.77 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from add [rotateZ(40deg)] to add [rotateZ(60deg)] at (0.25) should be [rotateZ(65deg)] assert_equals: expected "matrix ( 0.42 , 0.91 , - 0.91 , 0.42 , 0 , 0 ) " but got "matrix ( 0.71 , 0.71 , - 0.71 , 0.71 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from add [rotateZ(40deg)] to add [rotateZ(60deg)] at (0.5) should be [rotateZ(70deg)] assert_equals: expected "matrix ( 0.34 , 0.94 , - 0.94 , 0.34 , 0 , 0 ) " but got "matrix ( 0.64 , 0.77 , - 0.77 , 0.64 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from add [rotateZ(40deg)] to add [rotateZ(60deg)] at (0.75) should be [rotateZ(75deg)] assert_equals: expected "matrix ( 0.26 , 0.97 , - 0.97 , 0.26 , 0 , 0 ) " but got "matrix ( 0.57 , 0.82 , - 0.82 , 0.57 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from add [rotateZ(40deg)] to add [rotateZ(60deg)] at (1) should be [rotateZ(80deg)] assert_equals: expected "matrix ( 0.17 , 0.98 , - 0.98 , 0.17 , 0 , 0 ) " but got "matrix ( 0.5 , 0.87 , - 0.87 , 0.5 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from add [rotateZ(40deg)] to add [rotateZ(60deg)] at (1.5) should be [rotateZ(90deg)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 0 , 0 ) " but got "matrix ( 0.34 , 0.94 , - 0.94 , 0.34 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from add [rotate(45deg)] to add [rotate(225deg)] at (-0.5) should be [rotate(45deg) skew(10deg, 20deg) rotate(-45deg)] assert_equals: expected "matrix ( 0.73 , 0.09 , - 0.09 , 1.27 , 0 , 0 ) " but got "matrix ( 0.71 , - 0.71 , 0.71 , 0.71 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from add [rotate(45deg)] to add [rotate(225deg)] at (0) should be [rotate(45deg) skew(10deg, 20deg) rotate(45deg)] assert_equals: expected "matrix ( - 0.09 , 1.27 , - 0.73 , - 0.09 , 0 , 0 ) " but got "matrix ( 0.71 , 0.71 , - 0.71 , 0.71 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from add [rotate(45deg)] to add [rotate(225deg)] at (0.25) should be [rotate(45deg) skew(10deg, 20deg) rotate(90deg)] assert_equals: expected "matrix ( - 0.58 , 0.83 , - 0.45 , - 0.96 , 0 , 0 ) " but got "matrix ( 0 , 1 , - 1 , 0 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from add [rotate(45deg)] to add [rotate(225deg)] at (0.5) should be [rotate(45deg) skew(10deg, 20deg) rotate(135deg)] assert_equals: expected "matrix ( - 0.73 , - 0.09 , 0.09 , - 1.27 , 0 , 0 ) " but got "matrix ( - 0.71 , 0.71 , - 0.71 , - 0.71 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from add [rotate(45deg)] to add [rotate(225deg)] at (0.75) should be [rotate(45deg) skew(10deg, 20deg) rotate(180deg)] assert_equals: expected "matrix ( - 0.45 , - 0.96 , 0.58 , - 0.83 , 0 , 0 ) " but got "matrix ( - 1 , 0 , 0 , - 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from add [rotate(45deg)] to add [rotate(225deg)] at (1) should be [rotate(45deg) skew(10deg, 20deg) rotate(225deg)] assert_equals: expected "matrix ( 0.09 , - 1.27 , 0.73 , 0.09 , 0 , 0 ) " but got "matrix ( - 0.71 , - 0.71 , 0.71 , - 0.71 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from add [rotate(45deg)] to add [rotate(225deg)] at (1.5) should be [rotate(45deg) skew(10deg, 20deg) rotate(315deg)] assert_equals: expected "matrix ( 0.73 , 0.09 , - 0.09 , 1.27 , 0 , 0 ) " but got "matrix ( 0.71 , - 0.71 , 0.71 , 0.71 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateX(20deg)] from accumulate [rotateX(40deg)] to accumulate [rotateX(60deg)] at (-0.5) should be [rotateX(50deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.64 , 0.77 , 0 , 0 , - 0.77 , 0.64 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.87 , 0.5 , 0 , 0 , - 0.5 , 0.87 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(20deg)] from accumulate [rotateX(40deg)] to accumulate [rotateX(60deg)] at (0) should be [rotateX(60deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.5 , 0.87 , 0 , 0 , - 0.87 , 0.5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.77 , 0.64 , 0 , 0 , - 0.64 , 0.77 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(20deg)] from accumulate [rotateX(40deg)] to accumulate [rotateX(60deg)] at (0.25) should be [rotateX(65deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.42 , 0.91 , 0 , 0 , - 0.91 , 0.42 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.71 , 0.71 , 0 , 0 , - 0.71 , 0.71 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(20deg)] from accumulate [rotateX(40deg)] to accumulate [rotateX(60deg)] at (0.5) should be [rotateX(70deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.34 , 0.94 , 0 , 0 , - 0.94 , 0.34 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.64 , 0.77 , 0 , 0 , - 0.77 , 0.64 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(20deg)] from accumulate [rotateX(40deg)] to accumulate [rotateX(60deg)] at (0.75) should be [rotateX(75deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.26 , 0.97 , 0 , 0 , - 0.97 , 0.26 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.57 , 0.82 , 0 , 0 , - 0.82 , 0.57 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(20deg)] from accumulate [rotateX(40deg)] to accumulate [rotateX(60deg)] at (1) should be [rotateX(80deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.17 , 0.98 , 0 , 0 , - 0.98 , 0.17 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.5 , 0.87 , 0 , 0 , - 0.87 , 0.5 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(20deg)] from accumulate [rotateX(40deg)] to accumulate [rotateX(60deg)] at (1.5) should be [rotateX(90deg)] assert_equals: expected "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0 , 1 , 0 , 0 , - 1 , 0 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 1 , 0 , 0 , 0 , 0 , 0.34 , 0.94 , 0 , 0 , - 0.94 , 0.34 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from accumulate [rotateY(40deg)] to accumulate [rotateY(60deg)] at (-0.5) should be [rotateY(50deg)] assert_equals: expected "matrix3d ( 0.64 , 0 , - 0.77 , 0 , 0 , 1 , 0 , 0 , 0.77 , 0 , 0.64 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.87 , 0 , - 0.5 , 0 , 0 , 1 , 0 , 0 , 0.5 , 0 , 0.87 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from accumulate [rotateY(40deg)] to accumulate [rotateY(60deg)] at (0) should be [rotateY(60deg)] assert_equals: expected "matrix3d ( 0.5 , 0 , - 0.87 , 0 , 0 , 1 , 0 , 0 , 0.87 , 0 , 0.5 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.77 , 0 , - 0.64 , 0 , 0 , 1 , 0 , 0 , 0.64 , 0 , 0.77 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from accumulate [rotateY(40deg)] to accumulate [rotateY(60deg)] at (0.25) should be [rotateY(65deg)] assert_equals: expected "matrix3d ( 0.42 , 0 , - 0.91 , 0 , 0 , 1 , 0 , 0 , 0.91 , 0 , 0.42 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.71 , 0 , - 0.71 , 0 , 0 , 1 , 0 , 0 , 0.71 , 0 , 0.71 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from accumulate [rotateY(40deg)] to accumulate [rotateY(60deg)] at (0.5) should be [rotateY(70deg)] assert_equals: expected "matrix3d ( 0.34 , 0 , - 0.94 , 0 , 0 , 1 , 0 , 0 , 0.94 , 0 , 0.34 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.64 , 0 , - 0.77 , 0 , 0 , 1 , 0 , 0 , 0.77 , 0 , 0.64 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from accumulate [rotateY(40deg)] to accumulate [rotateY(60deg)] at (0.75) should be [rotateY(75deg)] assert_equals: expected "matrix3d ( 0.26 , 0 , - 0.97 , 0 , 0 , 1 , 0 , 0 , 0.97 , 0 , 0.26 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.57 , 0 , - 0.82 , 0 , 0 , 1 , 0 , 0 , 0.82 , 0 , 0.57 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from accumulate [rotateY(40deg)] to accumulate [rotateY(60deg)] at (1) should be [rotateY(80deg)] assert_equals: expected "matrix3d ( 0.17 , 0 , - 0.98 , 0 , 0 , 1 , 0 , 0 , 0.98 , 0 , 0.17 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.5 , 0 , - 0.87 , 0 , 0 , 1 , 0 , 0 , 0.87 , 0 , 0.5 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateY(20deg)] from accumulate [rotateY(40deg)] to accumulate [rotateY(60deg)] at (1.5) should be [rotateY(90deg)] assert_equals: expected "matrix3d ( 0 , 0 , - 1 , 0 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.34 , 0 , - 0.94 , 0 , 0 , 1 , 0 , 0 , 0.94 , 0 , 0.34 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from accumulate [rotateZ(40deg)] to accumulate [rotateZ(60deg)] at (-0.5) should be [rotateZ(50deg)] assert_equals: expected "matrix ( 0.64 , 0.77 , - 0.77 , 0.64 , 0 , 0 ) " but got "matrix ( 0.87 , 0.5 , - 0.5 , 0.87 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from accumulate [rotateZ(40deg)] to accumulate [rotateZ(60deg)] at (0) should be [rotateZ(60deg)] assert_equals: expected "matrix ( 0.5 , 0.87 , - 0.87 , 0.5 , 0 , 0 ) " but got "matrix ( 0.77 , 0.64 , - 0.64 , 0.77 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from accumulate [rotateZ(40deg)] to accumulate [rotateZ(60deg)] at (0.25) should be [rotateZ(65deg)] assert_equals: expected "matrix ( 0.42 , 0.91 , - 0.91 , 0.42 , 0 , 0 ) " but got "matrix ( 0.71 , 0.71 , - 0.71 , 0.71 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from accumulate [rotateZ(40deg)] to accumulate [rotateZ(60deg)] at (0.5) should be [rotateZ(70deg)] assert_equals: expected "matrix ( 0.34 , 0.94 , - 0.94 , 0.34 , 0 , 0 ) " but got "matrix ( 0.64 , 0.77 , - 0.77 , 0.64 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from accumulate [rotateZ(40deg)] to accumulate [rotateZ(60deg)] at (0.75) should be [rotateZ(75deg)] assert_equals: expected "matrix ( 0.26 , 0.97 , - 0.97 , 0.26 , 0 , 0 ) " but got "matrix ( 0.57 , 0.82 , - 0.82 , 0.57 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from accumulate [rotateZ(40deg)] to accumulate [rotateZ(60deg)] at (1) should be [rotateZ(80deg)] assert_equals: expected "matrix ( 0.17 , 0.98 , - 0.98 , 0.17 , 0 , 0 ) " but got "matrix ( 0.5 , 0.87 , - 0.87 , 0.5 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateZ(20deg)] from accumulate [rotateZ(40deg)] to accumulate [rotateZ(60deg)] at (1.5) should be [rotateZ(90deg)] assert_equals: expected "matrix ( 0 , 1 , - 1 , 0 , 0 , 0 ) " but got "matrix ( 0.34 , 0.94 , - 0.94 , 0.34 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from accumulate [rotateY(30deg)] to accumulate [rotateY(70deg)] at (-0.5) should be [rotateX(45deg) rotateY(10deg)] assert_equals: expected "matrix3d ( 0.98 , 0.12 , - 0.12 , 0 , 0 , 0.71 , 0.71 , 0 , 0.17 , - 0.7 , 0.7 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.98 , 0 , - 0.17 , 0 , 0 , 1 , 0 , 0 , 0.17 , 0 , 0.98 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from accumulate [rotateY(30deg)] to accumulate [rotateY(70deg)] at (0) should be [rotateX(45deg) rotateY(30deg)] assert_equals: expected "matrix3d ( 0.87 , 0.35 , - 0.35 , 0 , 0 , 0.71 , 0.71 , 0 , 0.5 , - 0.61 , 0.61 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.87 , 0 , - 0.5 , 0 , 0 , 1 , 0 , 0 , 0.5 , 0 , 0.87 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from accumulate [rotateY(30deg)] to accumulate [rotateY(70deg)] at (0.25) should be [rotateX(45deg) rotateY(40deg)] assert_equals: expected "matrix3d ( 0.77 , 0.45 , - 0.45 , 0 , 0 , 0.71 , 0.71 , 0 , 0.64 , - 0.54 , 0.54 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.77 , 0 , - 0.64 , 0 , 0 , 1 , 0 , 0 , 0.64 , 0 , 0.77 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from accumulate [rotateY(30deg)] to accumulate [rotateY(70deg)] at (0.5) should be [rotateX(45deg) rotateY(50deg)] assert_equals: expected "matrix3d ( 0.64 , 0.54 , - 0.54 , 0 , 0 , 0.71 , 0.71 , 0 , 0.77 , - 0.45 , 0.45 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.64 , 0 , - 0.77 , 0 , 0 , 1 , 0 , 0 , 0.77 , 0 , 0.64 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from accumulate [rotateY(30deg)] to accumulate [rotateY(70deg)] at (0.75) should be [rotateX(45deg) rotateY(60deg)] assert_equals: expected "matrix3d ( 0.5 , 0.61 , - 0.61 , 0 , 0 , 0.71 , 0.71 , 0 , 0.87 , - 0.35 , 0.35 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.5 , 0 , - 0.87 , 0 , 0 , 1 , 0 , 0 , 0.87 , 0 , 0.5 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from accumulate [rotateY(30deg)] to accumulate [rotateY(70deg)] at (1) should be [rotateX(45deg) rotateY(70deg)] assert_equals: expected "matrix3d ( 0.34 , 0.66 , - 0.66 , 0 , 0 , 0.71 , 0.71 , 0 , 0.94 , - 0.24 , 0.24 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0.34 , 0 , - 0.94 , 0 , 0 , 1 , 0 , 0 , 0.94 , 0 , 0.34 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotateX(45deg)] from accumulate [rotateY(30deg)] to accumulate [rotateY(70deg)] at (1.5) should be [rotateX(45deg) rotateY(90deg)] assert_equals: expected "matrix3d ( 0 , 0.71 , - 0.71 , 0 , 0 , 0.71 , 0.71 , 0 , 1 , 0 , 0 , 0 , 0 , 0 , 0 , 1 ) " but got "matrix3d ( 0 , 0 , - 1 , 0 , 0 , 1 , 0 , 0 , 1 , 0 , 0 , 0 , 0 , 0 , 0 , 1 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from accumulate [rotate(45deg)] to accumulate [rotate(225deg)] at (-0.5) should be [rotate(0deg) skew(10deg, 20deg)] assert_equals: expected "matrix ( 1 , 0.36 , 0.18 , 1 , 0 , 0 ) " but got "matrix ( 0.71 , - 0.71 , 0.71 , 0.71 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from accumulate [rotate(45deg)] to accumulate [rotate(225deg)] at (0) should be [rotate(90deg) skew(10deg, 20deg)] assert_equals: expected "matrix ( - 0.36 , 1 , - 1 , 0.18 , 0 , 0 ) " but got "matrix ( 0.71 , 0.71 , - 0.71 , 0.71 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from accumulate [rotate(45deg)] to accumulate [rotate(225deg)] at (0.25) should be [rotate(135deg) skew(10deg, 20deg)] assert_equals: expected "matrix ( - 0.96 , 0.45 , - 0.83 , - 0.58 , 0 , 0 ) " but got "matrix ( 0 , 1 , - 1 , 0 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from accumulate [rotate(45deg)] to accumulate [rotate(225deg)] at (0.5) should be [rotate(180deg) skew(10deg, 20deg)] assert_equals: expected "matrix ( - 1 , - 0.36 , - 0.18 , - 1 , 0 , 0 ) " but got "matrix ( - 0.71 , 0.71 , - 0.71 , - 0.71 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from accumulate [rotate(45deg)] to accumulate [rotate(225deg)] at (0.75) should be [rotate(225deg) skew(10deg, 20deg)] assert_equals: expected "matrix ( - 0.45 , - 0.96 , 0.58 , - 0.83 , 0 , 0 ) " but got "matrix ( - 1 , 0 , 0 , - 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from accumulate [rotate(45deg)] to accumulate [rotate(225deg)] at (1) should be [rotate(270deg) skew(10deg, 20deg)] assert_equals: expected "matrix ( 0.36 , - 1 , 1 , - 0.18 , 0 , 0 ) " but got "matrix ( - 0.71 , - 0.71 , 0.71 , - 0.71 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [rotate(45deg) skew(10deg, 20deg)] from accumulate [rotate(45deg)] to accumulate [rotate(225deg)] at (1.5) should be [rotate(360deg) skew(10deg, 20deg)] assert_equals: expected "matrix ( 1 , 0.36 , 0.18 , 1 , 0 , 0 ) " but got "matrix ( 0.71 , - 0.71 , 0.71 , 0.71 , 0 , 0 ) "
